feat: add network status

+++ app/src/main/java/com/example/mobileapplication/data/repository/RemoteBookRepository.kt
@@ -1,17 +1,23 @@
 package com.example.mobileapplication.data.repository
 
 import android.util.Log
+import com.example.mobileapplication.core.NetworkHelper
 import com.example.mobileapplication.data.remote.BookApiService
 import com.example.mobileapplication.domain.model.Book
 import com.example.mobileapplication.domain.repository.BookRepository
 
 class RemoteBookRepository(
-    private val apiService: BookApiService
+    private val apiService: BookApiService,
+    private val networkHelper: NetworkHelper
 ) : BookRepository {
 
     // 1. Добавить книгу на сервер
     override suspend fun insertBook(book: Book): Long {
         return try {
+            if (!networkHelper.isNetworkAvailable()) {
+                return 0
+            }
+
             val createdBook = apiService.postBook(book)
             // Возвращаем ID созданной сервером книги
             createdBook.id.toLong()
@@ -24,6 +30,9 @@ class RemoteBookRepository(
     // 2. Получить все книги напрямую с сервера
     override suspend fun getAllBooks(): List<Book> {
         return try {
+            if (!networkHelper.isNetworkAvailable()) {
+                return emptyList()
+            }
             apiService.getBooks()
         } catch (e: Exception) {
             Log.e("REPO", "Ошибка при получении списка: ${e.message}")
@@ -34,6 +43,9 @@ class RemoteBookRepository(
     // 3. Удалить книгу по ID на сервере
     override suspend fun deleteBook(id: Int): Int {
         return try {
+            if (!networkHelper.isNetworkAvailable()) {
+                return 0
+            }
             val response = apiService.deleteBook(id)
             if (response.isSuccessful) 1 else 0
         } catch (e: Exception) {
@@ -45,6 +57,9 @@ class RemoteBookRepository(
     // 5. Получить конкретную книгу по ID с сервера
     override suspend fun getBookById(id: Int): Book? {
         return try {
+            if (!networkHelper.isNetworkAvailable()) {
+                return null
+            }
             val response = apiService.getBook(id)
             if (response.isSuccessful) response.body() else null
         } catch (e: Exception) {
@@ -56,6 +71,9 @@ class RemoteBookRepository(
     // 6. Обновить книгу на сервере
     override suspend fun updateBook(book: Book): Int {
         return try {
+            if (!networkHelper.isNetworkAvailable()) {
+                return 0
+            }
             val response = apiService.putBook(book.id, book)
             // Если код ответа 200-299, считаем что обновили 1 запись
             if (response.isSuccessful) 1 else 0
